A bunch of work toward enabling self-healing
Allow users to create the cult ACL if it's not in place. This will
still fail if they haven't given us permissions.

While I was fixing this I made two relatively large changes:

1) start caching virtual docs in the model
2) tweak the reloading logic (rather than nulling the promise, just
put the tripledoc document promise in place)

little concerned this will lead to bugs, so let's test it!